requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
